# openapi/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Gestionale Negozio API
  description: API del gestionale negozio
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/interventions/{id}:
    get:
      tags:
      - intervention-controller
      operationId: getInterventionById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterventionResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      tags:
      - intervention-controller
      operationId: updateIntervention
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInterventionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterventionResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      tags:
      - intervention-controller
      operationId: deleteIntervention
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/interventions:
    get:
      tags:
      - intervention-controller
      operationId: getAllInterventions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InterventionResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      tags:
      - intervention-controller
      operationId: createIntervention
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInterventionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterventionResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/interventions/search:
    get:
      tags:
      - intervention-controller
      operationId: searchIntervention
      parameters:
      - name: nome
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InterventionResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/interventions/by-name/{nome}:
    get:
      tags:
      - intervention-controller
      operationId: getInterventionByName
      parameters:
      - name: nome
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterventionResponse"
        "404":
          description: Risorsa non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Non autorizzato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non autenticato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Richiesta non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Conflitto - risorsa già esistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
    ApiError:
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    InterventionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
          minLength: 1
        prezzo:
          type: number
        periodoGaranzia:
          type: integer
          format: int32
        cumulabile:
          type: boolean
      required:
      - id
      - nome
      - prezzo
...
    UpdateInterventionRequest:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        prezzo:
          type: number
        periodoGaranzia:
          type: integer
          format: int32
        cumulabile:
          type: boolean
      required:
      - nome
      - prezzo
...
    CreateInterventionRequest:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        prezzo:
          type: number
        periodoGaranzia:
          type: integer
          format: int32
        cumulabile:
          type: boolean
      required:
      - nome
      - prezzo
...